Phase 2: Extract UI event handler setup functions
Extract setupTooltips and attachWorkspaceEventListeners from main.ts
into a focused ui-event-handlers.ts module.

**What Was Extracted**:
- setupTooltips(): 25 lines - Setup tooltips for data-tooltip attributes
- attachWorkspaceEventListeners(): 49 lines - Workspace input/button event setup

**Impact**:
- main.ts: Reduced by 74 lines
- Created: src/lib/ui-event-handlers.ts (96 lines)
- Risk: Zero (pure event handler setup functions, no state mutation)
- Testing: TypeScript type check + Biome lint + Vite build all pass

**Benefits**:
- Clearer separation of concerns
- Easier to test event handler setup in isolation
- Continued incremental progress on issue #237

Issue #237
